Refactor refund

+++ app/Http/Controllers/Business/CustomerController.php
@@ -5,9 +5,14 @@ namespace App\Http\Controllers\Business;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Auth;
-
+use Maatwebsite\Excel\HeadingRowImport;
 use App\Customer;
 use App\StripePaymentMethod;
+use Illuminate\Support\Facades\Storage;
+use App\{ExcelUploadTracker};
+use Excel;
+use App\Imports\{CustomerImport,ImportMembership,customerAtendanceImport};
+use App\Jobs\{ProcessAttendanceExcelData,ProcessCustomerExcelData,ProcessMembershipExcelData};
 
 class CustomerController extends Controller
 {
@@ -98,36 +103,23 @@ class CustomerController extends Controller
         //
     }
 
-
-    public function visit_membership_modal(Request $request, $business_id ){
-        $user = Auth::user();
-        $company = $user->businesses()->findOrFail($business_id);
-        $customer = $company->customers->find($request->id);
-         $booking_detail = $customer->bookingDetail()->findOrFail($request->booking_detail_id);
-       // $booking_status = $customer->bookingStatus()->findOrFail($request->booking_id);
-        //$booking_detail = $booking_status->UserBookingDetail()->findOrFail($request->booking_detail_id);
-        return view('customers._edit_membership_info_model', ['booking_detail' => $booking_detail ,'business_id' =>$business_id ,"customer_id"=>$request->id]);
-    }
-
     public function card_editing_form(Request $request, $business_id){
         $company = $request->current_company;
-
         $customer = $company->customers()->findOrFail($request->customer_id);
         $stripe = new \Stripe\StripeClient(config('constants.STRIPE_KEY'));
-
+        $customer->create_stripe_customer_id();
         $intent = $stripe->setupIntents->create(
           [
-            'customer' => $customer->stripe_customer_id,
+            'customer' => @$customer->stripe_customer_id,
             'payment_method_types' => ['card'],
           ]
         );
-
         return view('business.customers.card_editing_form', compact('intent'));
     }
 
 
     public function refresh_payment_methods(Request $request){
-
+        
         $customer = Customer::findOrFail($request->customer_id);
         $stripe = new \Stripe\StripeClient(config('constants.STRIPE_KEY'));
         $payment_methods = $stripe->paymentMethods->all(['customer' => $customer->stripe_customer_id, 'type' => 'card']);
@@ -136,8 +128,10 @@ class CustomerController extends Controller
         foreach($payment_methods as $payment_method){
             $fingerprint = $payment_method['card']['fingerprint'];
             if (in_array($fingerprint, $fingerprints, true)) {
-                $deletePaymentMethod = StripePaymentMethod::where('payment_id', $payment_method['id'])->firstOrFail();
-                $deletePaymentMethod->delete();
+                $deletePaymentMethod = StripePaymentMethod::where('payment_id', $payment_method['id'])->first();
+                if($deletePaymentMethod != ''){
+                    $deletePaymentMethod->delete();
+                }
             } else {
                 $fingerprints[] = $fingerprint;
                 $stripePaymentMethod = StripePaymentMethod::firstOrNew([
@@ -155,7 +149,60 @@ class CustomerController extends Controller
             }
         }
         // echo $request->return_url;exit;
-        if($request->return_url)
+        if($request->return_url){
             return redirect($request->return_url);
+        }else{
+            return redirect()->route('business_customer_show',['business_id' => $customer->business_id , 'id' =>$request->customer_id ]);
+        }
+    }
+
+    public function importcustomer(Request $request, $business_id)
+    {
+        $user = Auth::user();
+        $current_company = $user->businesses()->findOrFail($business_id);
+        $file = $request->file('import_file');
+
+        // if($current_company->customer_uploading){
+        //    return response()->json(['status'=>500,'message'=>'Client import processing...']); 
+        // }
+        if($request->hasFile('import_file')){
+            $ext = $file->getClientOriginalExtension();
+            // if($ext != 'csv' && $ext != 'csvx' && $ext != 'xls' && $ext != 'xlsx' ){
+            if($ext != 'csv'){
+                return response()->json(['status'=>500,'message'=>'File format is not supported.']);
+            }
+
+            $headings = (new HeadingRowImport)->toArray($file);
+            /*print_r($headings);*/
+            if(!empty($headings)){
+                foreach($headings as $key => $row) {
+                    $firstrow = $row[0];
+                    /*print_r($firstrow);exit;*/
+                    if(  $firstrow[0] != 'last_name' || $firstrow[1] != 'first_name' || $firstrow[3] != 'id' ||  $firstrow[4] != 'address'|| $firstrow[5] != 'city'|| $firstrow[6] != 'state' || $firstrow[7] != 'postal_code' || $firstrow[8] != 'country'|| $firstrow[9] != 'mobile_phone' || $firstrow[10] != 'home_phone' || $firstrow[11] != 'work_phone' || $firstrow[12] != 'email') 
+                    {
+                        return response()->json(['status'=>500,'message'=>'Problem in header.']);
+                    }
+                }
+            }
+
+
+            $current_company->update(['customer_uploading' => 1]);
+            
+            
+            $timestamp = now()->timestamp;
+            $newFileName = 'business/customer_imports/'.$request->business_id.'-'.$timestamp.'-'.$user->id.'.'.$file->getClientOriginalExtension();
+
+
+            Storage::disk('s3')->put($newFileName, file_get_contents($file));
+
+
+            ProcessCustomerExcelData::dispatch($request->business_id, $newFileName);
+
+            
+
+            return response()->json(['status'=>200,'message'=>'File processing...']);
+        }
+
+
     }
 }
